# perf: use fxhash for `<=size_of::<usize>()`