--- a/src/lib.rs
+++ b/src/lib.rs
@@ -3,8 +3,8 @@
 // TODO: Better types & API please
 
 use ahash::AHashMap;
-
 use revm::primitives::{AccountInfo, Address, U256};
+use rustc_hash::FxHashMap;
 
 // TODO: More granularity here, for instance, to separate an account's
 // balance, nonce, etc. instead of marking conflict at the whole account.
@@ -81,7 +81,7 @@ type TransactionsStatus = Vec<TxIncarnationStatus>;
 // while adding new dependencies.
 // TODO: Intuitively both should share a smiliar data structure?
 type TransactionsDependents = Vec<Vec<TxIdx>>;
-type TransactionsDependencies = AHashMap<TxIdx, Vec<TxIdx>>;
+type TransactionsDependencies = FxHashMap<TxIdx, Vec<TxIdx>>;
 
 // BlockSTM maintains an in-memory multi-version data structure that
 // stores for each memory location the latest value written per