fxhash should be 10-20% faster than ahash for keys smaller than machine word size.
I also removed the comment about a different hasher, I believe it's not necessary for internal maps.

I would appreciate if you could run the benchmarks on this PR.

Very simple benchmarks with just `BuildHasher::hash_one` on x86_64 linux znver4:

```
u8
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 3.935691ms (0ns * 10000000)
                                  ahash::random_state::RandomState: 5.133483ms (1ns * 10000000)
                                    std::hash::random::RandomState: 54.898574ms (5ns * 10000000)
usize
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 4.04741ms (0ns * 10000000)
                                  ahash::random_state::RandomState: 4.798025ms (0ns * 10000000)
                                    std::hash::random::RandomState: 62.855416ms (6ns * 10000000)
[u8; 20]
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 7.719744ms (1ns * 10000000)
                                  ahash::random_state::RandomState: 6.722812ms (1ns * 10000000)
                                    std::hash::random::RandomState: 98.453998ms (10ns * 10000000)
[u8; 32]
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 12.286601ms (1ns * 10000000)
                                  ahash::random_state::RandomState: 7.480916ms (1ns * 10000000)
                                    std::hash::random::RandomState: 120.274741ms (12ns * 10000000)
[u64; 4]
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 12.322131ms (1ns * 10000000)
                                  ahash::random_state::RandomState: 7.427987ms (1ns * 10000000)
                                    std::hash::random::RandomState: 119.809424ms (12ns * 10000000)
bench_hashers::MemoryLocation
              core::hash::BuildHasherDefault<rustc_hash::FxHasher>: 29.677965ms (3ns * 10000000)
                                  ahash::random_state::RandomState: 21.344376ms (2ns * 10000000)
                                    std::hash::random::RandomState: 249.680294ms (25ns * 10000000)
```