@hai-rise 2024-06-11
@DaniPopes I ran the benchmark and the difference is not noticeable. I had a local branch that went further with an identity hasher for `TxIdx` that was also not good enough 🥲 . 

I think the main reason is that `transactions_dependencies` isn't usually hot. For hashing, `MemoryLocation` is much more interesting as it's the key to the multi-version data map (top 3 bottlenecks). It's also highly structured: a random hash with an optional big number. Would love to work with cryptographers on that eventually!

In the meantime, really love your code cleanups. Much more readable! Perhaps you can turn this one into another `chore`/`refactor` 🙏 ?